+++ CHANGES.md
@@ -3,9 +3,11 @@ Changes in Element 1.0.12 (2020-XX-XX)
 
 Features ✨:
  - Add room aliases management, and room directory visibility management in a dedicated screen (#1579, #2428)
+ - Room setting: update join rules and guest access (#2442)
 
 Improvements 🙌:
  - Add Setting Item to Change PIN (#2462)
+ - Improve room history visibility setting UX (#1579)
 
 Bugfix 🐛:
  - Double bottomsheet effect after verify with passphrase